Refactoring, commenting, clean up
