+++ core_java/twitter/src/main/java/ca/jrvs/apps/twitter/impl/TwitterController.java
@@ -1,4 +1,4 @@
-package ca.jrvs.apps.twitter;
+package ca.jrvs.apps.twitter.impl;
 
 import ca.jrvs.apps.twitter.controller.Controller;
 import ca.jrvs.apps.twitter.model.Tweet;
@@ -13,11 +13,18 @@ public class TwitterController implements Controller {
 
   private Service service;
 
+  /**
+   * The constructor sets up TwitterController object with
+   * Service objects, most likely TwitterService objects.
+   *
+   * @param service
+   */
   public TwitterController(Service service){this.service = service;}
 
   @Override
   public Tweet postTweet(String[] args) {
 
+    //Check if enough arguments provided
     if (args.length != 3){
       throw new IllegalArgumentException("Usage: TwitterAppCLI post \"text\" \"longitude:latitude\"");
     }
@@ -27,6 +34,7 @@ public class TwitterController implements Controller {
 
     String[] coords_arr = coordinates.split(COORD_SEP);
 
+    //Check if both coordinates provided and text argument is not blank
     if (coords_arr.length != 2 || text.isEmpty()){
       throw new IllegalArgumentException("Please provide non-empty arguments.\nUsage:"
           + " TwitterAppCLI post \"text\" \"longitude:latitude\"");
@@ -35,6 +43,10 @@ public class TwitterController implements Controller {
     float lon;
     float lat;
 
+    /*
+      Convert the coordinates from String to Float.
+      Any problems during conversion likely signifies lat/lon arguments being incorrectly provided.
+     */
     try{
 
       lon = Float.parseFloat(coords_arr[0]);
@@ -52,12 +64,14 @@ public class TwitterController implements Controller {
   @Override
   public Tweet showTweet(String[] args) {
 
+    //Check if enough arguments provided
     if (args.length != 2){
       throw new IllegalArgumentException("Usage: TwitterAppCLI show \"id\"");
     }
 
     String id = args[1];
 
+    //Check if id argument is not blank
     if (id.isEmpty()){
       throw new IllegalArgumentException("Please provide non-empty id value.\nUsage: TwitterAppCLI show \"id\"");
     }
@@ -68,6 +82,7 @@ public class TwitterController implements Controller {
   @Override
   public List<Tweet> deleteTweet(String[] args) {
 
+    //Check if enough arguments provided
     if(args.length < 2){
       throw new IllegalArgumentException("Usage: TwitterAppCLI delete \"id\",...,\"id\"");
     }
@@ -75,6 +90,7 @@ public class TwitterController implements Controller {
     String comma_Ids = args[1];
     String[] ids = comma_Ids.split(COMMA);
 
+    // Check there is at least one id
     if(ids.length < 1 || ids[0].isEmpty()){
       throw new IllegalArgumentException("Please provide a list of ids separated by commas. (min. 1)\n"
           + "Usage: TwitterAppCLI delete \"id\",...,\"id\"");
