+++ core_java/twitter/src/test/java/ca/jrvs/apps/twitter/TwitterControllerTest.java
@@ -5,6 +5,10 @@ import static org.junit.Assert.*;
 import ca.jrvs.apps.twitter.controller.Controller;
 import ca.jrvs.apps.twitter.dao.CrdDao;
 import ca.jrvs.apps.twitter.dao.helper.HttpHelper;
+import ca.jrvs.apps.twitter.impl.TwitterController;
+import ca.jrvs.apps.twitter.impl.TwitterDao;
+import ca.jrvs.apps.twitter.impl.TwitterHttpHelper;
+import ca.jrvs.apps.twitter.impl.TwitterService;
 import ca.jrvs.apps.twitter.model.Tweet;
 import ca.jrvs.apps.twitter.service.Service;
 import com.google.gdata.util.common.base.PercentEscaper;
