+++ core_java/twitter/src/test/java/ca/jrvs/apps/twitter/TwitterControllerUnitTest.java
@@ -4,6 +4,7 @@ import static org.junit.Assert.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
+import ca.jrvs.apps.twitter.impl.TwitterController;
 import ca.jrvs.apps.twitter.model.Tweet;
 import ca.jrvs.apps.twitter.service.Service;
 import ca.jrvs.apps.twitter.util.TweetBuilder;
